avs_lab5-main/requirements.txt
Flask==2.3.3
torch==2.0.1
torchvision==0.15.2
tqdm
Pillow==10.0.0
psycopg2-binary==2.9.7
boto3==1.28.30
gradio==3.41.2
numpy==1.24.3
requests==2.31.0
